+++ src/client/src/router/index.js
@@ -8,6 +8,7 @@ import AdminGroups from '../views/AdminGroups.vue'
 import AdminLogs from '../views/AdminLogs.vue'
 import ProxmoxManager from '../views/ProxmoxManager.vue'
 import MonitoringManager from '../views/MonitoringManager.vue'
+import Inventory from '../views/Inventory.vue'
 import ServerDetail from '../views/ServerDetail.vue'
 import ChangePasswordModal from '../components/ChangePasswordModal.vue'
 
@@ -56,6 +57,12 @@ const routes = [
     path: '/admin/logs',
     name: 'AdminLogs',
     component: AdminLogs,
+    meta: { requiresAuth: true, requiresAdmin: true }
+  },
+  {
+    path: '/admin/inventory',
+    name: 'Inventory',
+    component: Inventory,
     meta: { requiresAuth: true, requiresAdmin: true }
   },
   {